requirements.txt
fastapi
uvicorn
pandas
numpy
xgboost
pydantic
requests
python-dotenv
scikit-learn